+++ CHANGELOG.md
@@ -186,6 +186,7 @@
 - [9398](https://github.com/vegaprotocol/vega/issues/9398) - Fix division by zero panic in market liquidity
 - [9413](https://github.com/vegaprotocol/vega/issues/9413) - Fix range validation for SLA parameters
 - [9332](https://github.com/vegaprotocol/vega/issues/9332) - Ethereum oracles sending data to unintended destinations
+- [9447](https://github.com/vegaprotocol/vega/issues/9447) - Store current SLA parameters in the liquidity engine snapshot.
 - [9433](https://github.com/vegaprotocol/vega/issues/9433) - fix referral set snapshot
 - [9432](https://github.com/vegaprotocol/vega/issues/9432) - fix referral set not saved to database.
 - [9449](https://github.com/vegaprotocol/vega/issues/9449) - if expiration is empty, never expire a discount/reward program